+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Calendar Versioning](https://calver.org/) of
 the following form: YYYY.0M.0D.
 
+## [2025.11.06]
+
+### Added
+
+- Created ChampionDBSecondary, including hashcode and equals
+- fixed syntax issues with implementations
+- added javadoc for all methods in abstract class
+
 
 ## [2025.10.23]